Moved Dao to database folder
